+++ header.js
@@ -93,26 +93,24 @@ BlockHeader.prototype.canonicalDifficulty = function (parentBlock) {
   var offset = parentDif.div(new BN(params.difficultyBoundDivisor.v))
   var dif
 
-  if (this.isHomestead()) {
-    // homestead
-    // 1 - (block_timestamp - parent_timestamp) // 10
-    var a = blockTs.sub(parentTs).idivn(10).ineg().iaddn(1)
-    var cutoff = new BN(-99)
-    // MAX(cutoff, a)
-    if (cutoff.cmp(a) === 1) {
-      a = cutoff
-    }
-    dif = parentDif.add(offset.mul(a))
-  } else {
-    // prehomestead
-    if (parentTs.addn(params.durationLimit.v).cmp(blockTs) === 1) {
-      dif = offset.add(parentDif)
-    } else {
-      dif = parentDif.sub(offset)
-    }
+  // Byzantium
+  // max((2 if len(parent.uncles) else 1) - ((timestamp - parent.timestamp) // 9), -99)
+  var uncleAddend = parentBlock.header.uncleHash.equals(utils.SHA3_RLP_ARRAY) ? 1 : 2
+  var a = blockTs.sub(parentTs).idivn(9).ineg().iaddn(uncleAddend)
+  var cutoff = new BN(-99)
+  // MAX(cutoff, a)
+  if (cutoff.cmp(a) === 1) {
+    a = cutoff
+  }
+  dif = parentDif.add(offset.mul(a))
+
+  // Byzantium difficulty bomb delay
+  var num = new BN(this.number).isubn(3000000)
+  if (num.ltn(0)) {
+    num = new BN(0)
   }
 
-  var exp = new BN(this.number).idivn(100000).isubn(2)
+  var exp = num.idivn(100000).isubn(2)
   if (!exp.isNeg()) {
     dif.iadd(new BN(2).pow(exp))
   }
@@ -231,20 +229,3 @@ BlockHeader.prototype.isGenesis = function () {
   return this.number.toString('hex') === ''
 }
 
-/**
- * Determines if a given block part of homestead or not
- * @method isHomestead
- * @return Boolean
- */
-BlockHeader.prototype.isHomestead = function () {
-  return utils.bufferToInt(this.number) >= params.homeSteadForkNumber.v
-}
-
-/**
- * Determines if a given block part of Homestead Reprice (EIP150) or not
- * @method isHomesteadReprice
- * @return Boolean
- */
-BlockHeader.prototype.isHomesteadReprice = function () {
-  return utils.bufferToInt(this.number) >= params.homesteadRepriceForkNumber.v
-}