+++ CHANGELOG.md
@@ -6,8 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
-## [1.7.0] - Unreleased
-- TODO
+## [1.7.0] - Unreleased (``master`` branch)
+- ``Metro-Byzantium`` compatible
+- New difficulty formula (EIP 100)
+- Difficulty bomb delay (EIP 649)
+- Removed ``isHomestead``, ``isHomesteadReprice`` from API methods
 
 [1.7.0]: https://github.com/ethereumjs/ethereumjs-vm/compare/v1.6.0...v1.7.0